requirements/base.txt
django
django-nose
djangorestframework
django-braces
django-grappelli
django-vanilla-views
django-cors-headers
